move backend related utility functions to nilearn.plotting.matrix._backend
